+++ test/truffle/cexts/puma/puma.sh
@@ -4,7 +4,6 @@ source test/truffle/common.sh.inc
 
 jt gem-test-pack
 
-export TRUFFLERUBY_CEXT_ENABLED=true
 jt ruby -S gem install --local "$(jt gem-test-pack)/gem-cache/puma-3.10.0.gem" -V -N --backtrace
 
 jt ruby "-I$(jt gem-test-pack)/gems/gems/rack-1.6.1/lib" -S \