+++ spec/truffle/interop/sizep_spec.rb
@@ -14,8 +14,8 @@ describe "Truffle::Interop.size?" do
     Truffle::Interop.size?([]).should be_true
   end
   
-  it "returns true for hashes" do
-    Truffle::Interop.size?({}).should be_true
+  it "returns false for hashes" do
+    Truffle::Interop.size?({}).should be_false
   end
   
   it "returns false for strings" do