+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Changes:
 * The `KEYS` interop message now returns an array of Java strings, rather than
   Ruby strings. `KEYS` on an array no longer returns indices.
 
+* A method call on a foreign object that looks like an operator (the method name
+  does not begin with a letter) will call `IS_BOXED` on the object and based on
+  that will possibly `UNBOX` and convert to Ruby.
+
 # 0.33, April 2018
 
 New features: