+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Changes:
 * The `KEYS` interop message now returns an array of Java strings, rather than
   Ruby strings. `KEYS` on an array no longer returns indices.
   
-* `HAS_SIZE` now returns `false` for `String`.
+* `HAS_SIZE` now only return `true` for `Array`.
 
 * A method call on a foreign object that looks like an operator (the method name
   does not begin with a letter) will call `IS_BOXED` on the object and based on